[WIP] Discard localizations global variable and use selection field instead
Because the localizations variables are set in res.company model
and it is a field, it should be an odoo selection field.
There is a method to recover the list (value, label) from a field
in a certain model, but it gives me an error that I can figure out
how to resolve it. I'll create an issue fo this matter