[REL] Merge pr from Bmya (#86)
* [WIP] First working version in odoo 11
Related to issue #4

Remember that some lines in report_invoice.xml were hidden until we finally agree
if new lines in odoo are the same

* [MIG] account_document: Replace openerp imports to odoo
Related to Issue #5

* [MIG] account_document: Rename manifest
Related to Issue #5

* [WIP] Discard localizations global variable and use selection field instead

Because the localizations variables are set in res.company model
and it is a field, it should be an odoo selection field.
There is a method to recover the list (value, label) from a field
in a certain model, but it gives me an error that I can figure out
how to resolve it. I'll create an issue fo this matter

* [FIX] account_document: Fix error in res.config
This error was related to the instalation of l10n_cl

No longer exist the model ir.values in v11

To get default values it is used ir.default instead

* [ADD] account_document: add base menu for tax entities

This is necessary for future adds like chilean sii or argentinian afip
in their corresponding localizations

* [IMP] account_document: Include vals in update_journal_document_types parameters to receive wizard vars in l10n_cl_account

* [FIX] account_document: Fix domain for selection of document types in invoice
depending on availables client document types

* [FIX] account_document: Remove constraint which forced update_journal_document_types() method being executed at create

This was causing to search for all available document types
regarding of the inherit method called by configuration wizard

Because there was already loaded a several document types into the o2m field
the new search didn't affect it. The algorithm doesn't remove lines, just add
the new ones.

* [FIX] add state = code on cron data

* [FIX] [WIP] account_document: Comment redefeinition of _get_tax_amount_by_group

It is causing a critical error at invoice print because it returns a wrong value
We have to fix the method according to the real need

* [FIX] WARNING odoo11a1 odoo.fields: Field account.payment.display_name depends on itself; please fix its decorator @api.depends().

* [REF] minor modifications

* Revert "[FIX] account_document: Remove constraint which forced update_journal_document_types() method being executed at create"

This reverts commit 5ea7444b85b5bba11ec43d09859dc59c41808474.

* Revert "[IMP] account_document: Include vals in update_journal_document_types parameters to receive wizard vars in l10n_cl_account"

This reverts commit 8ad1f3411927afdad6b1b2ec7f8c3e5184189560.

* [FIX] ADD repo dependencies