[FIX] account_interests: proper set account/taxes in lines of interest invoices
Formerly invoice lines were not taking into account the product config account
and taxes, this was happending because an onchange that requires that the
invoice exists before in order to proper compute this data. Taking this into
account now we create first the invoice and then we add and compute the
invoice lines with the proper onchanges.