[FIX] Port-foward from 9.0 (#125)
* [FIX] account_interests: Multiple fixes for this module (#124)

* [FIX] account_interests: Proper date format in Due Since message

Formerly was using the database string date format, now using the lang
configured in the context or the current user in order to print the date
properly.

* [IMP] account_interests: Show interest product name as invoice name

Formerly we where using reference but this field is not used anymore.

* [IMP] account_interests: Account validation not needed

The account used for the new interest invoice is related to the account
configured in the interest product: if this one is not set then will just
raise a user error asking for this proper configuration

* [FIX] account_interests: proper set account/taxes in lines of interest
invoices

Formerly invoice lines were not taking into account the product config account
and taxes, this was happending because an onchange that requires that the
invoice exists before in order to proper compute this data. Taking this into
account now we create first the invoice and then we add and compute the
invoice lines with the proper onchanges.

* [FIX] account_interests: Odoo v9 deprecate analytic account type field.

This one was used to filter the analytic account to be set from the account
interest configurator.

Now we filter by the accounts that are active, do not show the close accounts.

* [ADD] account_interests: Just update module version

To be able to updagre this module with all the new changes in the PR.

* [FIX] account_interests: account_type field deprecated

Analytic account field in the interest configurator has a domain that use an
account_type  which is deprecated. version 11.0 now use active field instead
so this domain is not needed anymore.  Odoo will filter the active registers
automatically.

* [FIX] account_interests: 11.0 _lang_get method return a recordset directly.
not an id.

* [IMP] account_interests: directly create invoice line since this one already
has invoice_id