[ADD] account_ux: do not let the user create rates without company if already had rates with company before the new rate date
This is required since create new records without company will make then
invisible, odoo will search for the ones with the last date with company
set. If a user create a rate without company this one will never be take into
account. Adding the constraint we avoid this human error.