[FIX] account_document: revert change that cause error
Currently odoo server is throwing an error that do not let to select a
contrapart in the conciliation widget.

This is was cause becase the method firm was change in order to try to
overwrite the str variable type, but some reasons this fails.

I think is because is not a regular inherit and need to have the same
name of the arguments.