[FIX] accout_document: Change doc types
If you create an advanced invoice and the journal changes with the company changes the documentes type are not update to the corresponding for the new journal. To fix that issue, if you chage the journal, the documents type are changes too. like the onchange do at the form view.