[ADD] account_ux: invoice-journal-type constrains
You can only create sales/purchase invoices in sale/purhcase journals.
We found that is a common user error when the user defind some defaults
wrongly. This fix is to avoid the user to create an customer invoices in
a purchase journal or any other journal different from sale. Also apply
for vendor bills.