[FIX]account_ux: compute currency rate
The calculation of the quote depends on the date of the invoice, however this date must be different for the type of invoice:

If it is a customer invoice we should take invoice_date
If it is a supplier invoice we should take date (accounting date).
This change was necessary because we get the following case: in the case of customer invoices, if we take the date field, what it will do is calculate the quote with the date on which the invoice was created. And this is a problem because if the user creates them before and validates them some time later, they will always take into account an old quote, taking into account that the date field is not shown in customer invoices and therefore they cannot edit.