[ADD] account_ux: Fix inherit view
We were trying to modify the debit_origin_id field for the original view
of account.move and not from actually the account_debit_note view that
is actually adding the field.

In this case, Odoo resolves it and does not show any warning or error
when loading the view, but we need this fix because this is causing an
error when installing or updating the module.

closes ingadhoc/account-financial-tools#334

X-original-commit: 9a98c2d3004daf1bb744df6bb95a94cc5fd9099c
Signed-off-by: Katherine Zaoral <kz@adhoc.com.ar>