[FIX] account_payment_term_surcharge: Correct calculation of next_surcharge_date with day_current_month
This fix addresses an issue when calculating `next_surcharge_date` for surcharges that depend on a specific day of the current month. For example, if the invoice date is 30/11 and the surcharge is set for the 13th of the month, the calculated `next_surcharge_date` was incorrectly set to 13/11 (a past date).

With this fix, the calculation ensures the `next_surcharge_date` is properly set to 13/12 (the next valid occurrence).

X-original-commit: bed924db913ba779238fb93adb4d9cf0adb15fd7