[FIX] account_journal_security: Fix access rights when having journal_restriction in modification mode.
When having journal_restriction in modification mode and trying to create a new account move a raise was raised
because was inconsistent the domain.

If journal_ids variable is defined, it's not necessary (nor valid) to explicitly set modification_journal_ids to False.
In fact, with journal_restriction enabled, it's not possible for modification_journal_ids to be False, as this would contradict the restriction logic.

closes ingadhoc/account-financial-tools#740

Signed-off-by: Filoquin adhoc <maq@adhoc.com.ar>