[FIX] account_journal_security: Fix access rights when having journal_restriction in modification mode.
When having journal_restriction in modification mode and trying to create a new account move a raise was raised
because was inconsistent the domain.

If journal_ids variable is defined, it's not necessary (nor valid) to explicitly set modification_journal_ids to False.
In fact, with journal_restriction enabled, it's not possible for modification_journal_ids to be False, as this would contradict the restriction logic.

closes ingadhoc/account-financial-tools#741

X-original-commit: 5e6e3d5a0a5d7fdd45983a24e80b59a4d3fd7fb1
Signed-off-by: Filoquin adhoc <maq@adhoc.com.ar>
Signed-off-by: rov-adhoc <rov@adhoc.com.ar>