[FIX] account_interest: fix safe_eval context type to avoid warning
This commit fixes a warning raised by safe_eval due to an incorrect evaluation context type.

The _get_eval_context method was returning a tuple instead of a dictionary

We now return a plain dictionary to ensure the evaluation context is correctly interpreted, avoiding the warning and ensuring compatibility with future Odoo versions.

closes ingadhoc/account-financial-tools#761

X-original-commit: 10b1a6daa20b7059e77cf4c557e0da6d325b88ee
Signed-off-by: Camila Vives <cav@adhoc.com.ar>
Signed-off-by: rov-adhoc <rov@adhoc.com.ar>