[FIX] account_debt_report: Fix inconsistent _read_group result handling in debt report
Ensure consistent handling of _read_group output when computing initial balance in the debt report.

Odoo's _read_group may return either a list of tuples or dictionaries depending on internal usage or parameters.
This patch checks the type of the result safely and avoids errors by validating the presence of data before accessing indices.

Also prevents a TypeError when the result is empty or has unexpected structure.

closes ingadhoc/account-financial-tools#760

X-original-commit: 177563a6f154f94c5043497b2f7cae439ee3ebe2
Signed-off-by: Felipe Garcia Suez <feg@adhoc.com.ar>
Signed-off-by: rov-adhoc <rov@adhoc.com.ar>