[IMP] account_ux: Unlink inactive tax when duplicate invoice
When archiving/deactivating taxes that were used in OS and invoices and then duplicating any of these records, the system copies the
lines while keeping the archived taxes in the new record. This allows you to validate OV or invoice with archived taxes at the time
of execution. Ideally, it should allow duplication, but the idea is to delete the tax from the line if it is archived.

closes ingadhoc/account-financial-tools#820

X-original-commit: 0f0631151e17920b1d518cd53b56de0c92a3f144
Related: ingadhoc/sale#1457
Signed-off-by: rov-adhoc <rov@adhoc.com.ar>