[IMP] account_internal_transfer: transfer entries in company currency
In a transfer between a journal in main currency and one in secondary currency,
the amount field in the journal entry for the main currency journal must be adjusted
so it is recorded in the main currency, not in the secondary currency. This commit fixes that behavior.

Part-of: ingadhoc/account-financial-tools#852
Related: ingadhoc/account-payment#850
Signed-off-by: rov-adhoc <rov@adhoc.com.ar>