[FIX] _exchange_difference_invoice: Fix tax calculation with fiscal position
When calculating tax groups for proration in the exchange difference wizard, we need to ensure that
if the invoice has a fiscal position, the perceptions and other tax amounts should not be included in the subtotal used for proration.
This case only applies if the user selects automatic fiscal position application in the wizard.

Also, if user selects manual and not fiscal position id, the fiscal position will be cleaned from the invoice
to avoid re-calculation taxes when adding the date.

closes ingadhoc/account-financial-tools#851

Signed-off-by: Camila Vives <cav@adhoc.inc>